+++ core/dbt/parser/sources.py
@@ -10,6 +10,7 @@ from dbt.context.context_config import (
     ContextConfigGenerator,
     UnrenderedConfigGenerator,
 )
+from dbt.contracts.graph.components import Time
 from dbt.contracts.graph.manifest import Manifest, SourceKey
 from dbt.contracts.graph.model_config import SourceConfig
 from dbt.contracts.graph.nodes import (
@@ -24,7 +25,6 @@ from dbt.contracts.graph.unparsed import (
     UnparsedSourceTableDefinition,
     FreshnessThreshold,
     UnparsedColumn,
-    Time,
 )
 from dbt_common.events.functions import warn_or_error, fire_event
 from dbt.events.types import UnusedTables, FreshnessConfigProblem
