+++ tests/unit/test_contracts_graph_parsed.py
@@ -21,6 +21,7 @@ from dbt.artifacts.resources import (
 from dbt.artifacts.resources.types import TimePeriod
 from dbt.node_types import NodeType, AccessType
 from dbt.contracts.files import FileHash
+from dbt.contracts.graph.components import Time
 from dbt.contracts.graph.model_config import (
     ModelConfig,
     NodeConfig,
@@ -49,10 +50,7 @@ from dbt.contracts.graph.nodes import (
     TestMetadata,
     SemanticModel,
 )
-from dbt.contracts.graph.unparsed import (
-    FreshnessThreshold,
-    Time,
-)
+from dbt.contracts.graph.unparsed import FreshnessThreshold
 from dbt import flags
 from argparse import Namespace
